+++ components/esp_lvgl_port/include/esp_lvgl_port.h
@@ -62,7 +62,7 @@ typedef struct {
 #define ESP_LVGL_PORT_INIT_CONFIG() \
     {                               \
         .task_priority = 4,       \
-        .task_stack = 6144,       \
+        .task_stack = 7168,       \
         .task_affinity = -1,      \
         .task_max_sleep_ms = 500, \
         .timer_period_ms = 5,     \